TT-6828-9 feat: implement Phrase Back Translation and Guided Phrase Record features
Also addresses:
- TT-6828, TT-6829 — guided segmenting + autosave
- TT-7053, TT-7103, TT-7095, TT-7096, TT-7115 — Save Segments / autosave / unsaved-dialog class
- TT-7093 — no verse pre-fill; listen-first + auto-segment flow
- TT-7133 — region navigation + play (CS playback model)
- TT-7385 — speaker cleared on record (CS controls)
- TT-7097 — dual-player legacy bugs
- TT-7094 — mobile routing to new component
- TT-7116 — overwrite semantics (CS delete + re-record)

Also likely addressed:
- TT-4856 — BT playlist
- TT-7082, TT-7109, TT-7110 — layout/design deltas; re-check against new CS-style layout
- Added new segment definitions and boundaries for Phrase Back Translation in CONTEXT.md.
- Updated localization files to include new strings for Phrase Back Translation controls.
- Refactored PassageDetail components to integrate Guided Phrase Record functionality, replacing the previous Careful Speech implementation.
- Adjusted tests to accommodate changes in segment handling and control strings.
