+++ CONTEXT.md
@@ -168,9 +168,21 @@ _Avoid_: Next workflow step; chevron navigation without play
 During the **recording pass**, tapping an **already-recorded** clause replays the source clause and loads that clause's existing **Careful Speech recording** in the recorder (where it can be played). The user may delete it with the trash control and record again. Tapping an unrecorded clause highlights it yellow, snaps to the start, and plays it. After the clause plays through, the clause stays yellow, the play position snaps back to the clause start, and the **Record button becomes enabled** (the primary next action). The play button is not re-highlighted — the user can replay, but recording is the expected next step. If the clause is too short to make a clear careful-speech recording, the alternative is **Combine with Next Clause**; if too long, **Split Clause**. If the user navigates to a different clause (by tapping it on the waveform) and then returns, the system plays the clause again automatically — the user must hear it before Record re-enables. Navigating away forfeits the heard state for that clause.
 _Avoid_: Silently overwriting a recording; ignoring recorded clauses; re-highlighting the play button after play-through; advancing to the next clause automatically after play-through (listen-pass behavior); enabling Record without the user having heard the current clause in the current navigation
 
+**Segment** (Phrase Back Translation):
+A numbered unit on the vernacular **source recording** under the **`BT`** named region — one time range, one **PBT** (or **Retell** artifact on Retell Back Translation steps). User-facing label in segment navigation (e.g. "Segment 3/7"). Retell Back Translation uses a single segment spanning the full passage.
+_Avoid_: Clause (in Phrase Back Translation UI when **Segment** is shown); Passage; separate boundary copies per artifact
+
+**Segment boundaries**:
+The time ranges that divide the vernacular **source recording** into **segments** for **Phrase Back Translation** and related `phraseBackTranslate` steps. Stored on the vernacular media under **`BT`** (`NamedRegions.BackTranslation`) — the canonical copy for non-BOLD phrase-level back translation. Set and adjusted during the **listen pass**; locked in the **recording pass**. **PBT** and **Retell** recordings link to a segment by matching its time range via `sourceSegments`, not by storing their own boundaries.
+_Avoid_: Clause boundaries (BOLD **`clause`** bucket); Mark Verses regions; boundaries stored only on back-translation artifacts
+
+**Retell Back Translation** (step):
+A non-BOLD `phraseBackTranslate` step configured with the **Retell** artifact where the user retells the entire passage for the consultant in one guided **listen pass** / **recording pass** — exactly **one segment** covering full vernacular audio, with no More/Fewer/Split/Combine controls. Legacy recordings with empty `sourceSegments` still count as complete for that segment.
+_Avoid_: Community Testing Retell; multi-segment Retell BT; Careful Speech
+
 **Phrase Back Translation** (step):
-A non-BOLD workflow step where the user listens to full-speed **vernacular** by phrase and records the meaning as a **PBT** in **LWC** — one workflow step with no separate **Careful Speech** artifact step (e.g. Kom directly to English). Uses the same **listen pass** / **recording pass** guided interaction as BOLD **Careful Speech** (planned to replace the legacy phrase-back-translate UI) via the `phraseBackTranslate` tool. **User-facing name is Phrase Back Translation, not Careful Speech** — only the labels and artifact produced differ from BOLD (PBT in LWC, not careful speech in vernacular).
-_Avoid_: Careful Speech (name or step outside BOLD); LWC Translation (BOLD step); Whole Back Translation; calling the output careful speech; single-phase-only UX once migration is complete
+A non-BOLD workflow step where the user listens to full-speed **vernacular** by phrase and records the meaning as a **PBT** in **LWC** — one workflow step with no separate **Careful Speech** artifact step (e.g. Kom directly to English). Uses the **listen pass** / **recording pass** guided interaction via `PassageDetailPhraseBackTranslate` and the `phraseBackTranslate` tool. **User-facing name is Phrase Back Translation, not Careful Speech** — only the labels (**Segment** not Clause) and artifact produced differ from BOLD (PBT in LWC, not careful speech in vernacular).
+_Avoid_: Careful Speech (name or step outside BOLD); LWC Translation (BOLD step); Whole Back Translation; calling the output careful speech; legacy single-phase-only UX
 
 **Phrase Back Translate** (tool):
 Developer-facing tool slug for phrase-level listen-and-record steps on non-BOLD processes (**Phrase Back Translation**) and for **LWC Translation** in BOLD (PBT artifact, LWC audio). BOLD **Careful Speech** uses the separate `carefulSpeech` tool and **Careful Speech artifact** type — that tool and step name appear only in BOLD workflows. Configured per step with artifact type and named region in **StepEditor**.
