+++ docs/adr/0005-bold-careful-speech-two-phase-flow.md
@@ -2,7 +2,7 @@
 
 BOLD Careful Speech is a two-phase step on the source recording from Record — not a vernacular version fork. Output is stored as the **`carefulspeech`** artifact type (vernacular re-speaking per clause), not PBT. The step uses the dedicated **`carefulSpeech`** tool; LWC Translation remains on **`phraseBackTranslate`** with PBT artifacts.
 
-Non-BOLD **Phrase Back Translation** is one workflow step (vernacular → PBT in LWC, no Careful Speech artifact) but will use the same **listen pass** / **recording pass** UX when the legacy UI is replaced; only labels and artifact type differ.
+Non-BOLD **Phrase Back Translation** is one workflow step (vernacular → PBT in LWC, no Careful Speech artifact) but uses the same **listen pass** / **recording pass** UX via `PassageDetailPhraseBackTranslate` and `PassageDetailGuidedPhraseRecord`. Segment boundaries persist under the legacy **`BT`** named region on vernacular media. **Retell Back Translation** on the same tool uses single-segment mode (full passage, no boundary tools). Resume rules match Careful Speech; legacy Retell takes with empty `sourceSegments` count as complete for the single segment.
 
 **Listen pass:** Auto-segment into clauses (default silence threshold 2, minimum silence length 2, minimum segment length 1.5). User must hear every clause once before Start Recording is enabled. More Clauses / Fewer Clauses adjust segmentation parameters and reset listen progress. Longer / Shorter boundary nudging is not used in BOLD. The listen-pass panel includes an info message explaining More Clauses and Combine with Next Clause.
