+++ src/renderer/src/store/localization/exported-strings-name.json
@@ -1,1 +1,1 @@
-{ "stringsName": "strings8dbb98c7.json" }
+{ "stringsName": "strings3f97fb5c.json" }
